requirements.txt
django==2.0.3
django-jet
google-api-python-client==1.4.1
FeedParser
djangorestframework
django-ckeditor
drfdocs
numpy
Pillow
nltk
pdfminer3k
underthesea
langdetect
pandas
sklearn
opencv-python